Add Future class to the project.
This class will represent the result of a future computation
in our event-based system. It will make it possible for callers
to get the event ID of their call before the results are returned.
This event ID will allow plugins to filter events based on ID values
(making it easier to write certain kinds of plugin operations).