Add monitor notification functions to EventFiringFuture.
An event-firing class can use the notifyMonitors(Result) method
to notify any event monitoring methods of the outcome of an event.
This method will send the Result to all event monitors on separate
threads and returns an MonitorNotifyFuture to allow clients to
track the progress of the monitor notification process.